pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixelcore"
version = "0.1.0"
description = "Multi engine pixel graphics system"
authors = [
    { name = "Shloke" }
]

dependencies = [
    "numpy",
    "pillow",
    "pygame"
]